pyproject.toml
[project]
name = "hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=26.1.0",
    "isort>=8.0.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.10",
    "python-dotenv>=1.2.2",
]
